refactor: move srcset format impl
